pyproject.toml
[tool.poetry]
name = "covid19-aws-pipeline"
version = "0.1.0"
description = ""
authors = ["danielmschaves <danielmschaves@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
boto3 = "^1.34.64"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
duckdb = "^0.9.2"
dbt-duckdb = "^1.7.3"
awswrangler = "^3.7.3"
pandas = "^2.2.2"
pyathena = "^3.8.2"
dbt-redshift = "^1.8.0"
dbt-athena-community = "^1.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
